ditch the ts extension (too much overhead)
