--- force-app/main/default/lwc/surveyQuestionMapping/surveyQuestionMapping.css
@@ -1,8 +1,0 @@
-.mapping-container {
-    padding: 0.5rem;
-}
-
-.slds-box {
-    background-color: #f3f2f2;
-    border-radius: 4px;
-}
