--- force-app/main/default/lwc/surveyQuestionMapping/surveyQuestionMapping.html
@@ -1,33 +1,0 @@
-<template>
-    <div class="mapping-container">
-        <template if:true={isLoading}>
-            <lightning-spinner alternative-text="Loading" size="small"></lightning-spinner>
-        </template>
-
-        <template if:true={error}>
-            <div class="slds-text-color_error">{error}</div>
-        </template>
-
-        <div class="slds-grid slds-gutters">
-            <div class="slds-col">
-                <lightning-combobox label="Salesforce Object" value={selectedObject} options={objectOptions}
-                    onchange={handleObjectChange} placeholder="Select an object"></lightning-combobox>
-            </div>
-            <div class="slds-col">
-                <lightning-combobox label="Field" value={selectedField} options={fieldOptions}
-                    onchange={handleFieldChange} placeholder="Select a field"
-                    disabled={fieldSelectDisabled}></lightning-combobox>
-            </div>
-        </div>
-
-        <template if:true={selectedObject}>
-            <div class="slds-m-top_small">
-                <div class="slds-text-title_caps slds-m-bottom_x-small">Current Mapping:</div>
-                <div class="slds-box slds-box_small slds-theme_shade">
-                    <p><strong>Object:</strong> {selectedObject}</p>
-                    <p><strong>Field:</strong> {selectedField}</p>
-                </div>
-            </div>
-        </template>
-    </div>
-</template>
