+++ force-app/main/default/lwc/uuidDisplay/uuidDisplay.js-meta.xml
@@ -1,10 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <LightningComponentBundle xmlns="http://soap.sforce.com/2006/04/metadata">
     <apiVersion>62.0</apiVersion>
-    <isExposed>true</isExposed>
-    <targets>
-        <target>lightning__AppPage</target>
-        <target>lightning__RecordPage</target>
-        <target>lightning__HomePage</target>
-    </targets>
+    <isExposed>false</isExposed>
+    <description>Utility component that registers a UUID display property editor with SurveyJS</description>
 </LightningComponentBundle>
