--- force-app/main/default/staticresources/surveyExtensions/uuidPropertyEditor.js
@@ -1,1 +1,0 @@
-!function(e,o){"object"==typeof exports&&"object"==typeof module?module.exports=o():"function"==typeof define&&define.amd?define([],o):"object"==typeof exports?exports.uuidPropertyEditor=o():e.uuidPropertyEditor=o()}(this,(()=>(()=>{"use strict";var e={},o=function(){function e(){}return e.register=function(){if(console.log("Registering UUID property editor"),void 0!==window.Survey){var o=window.Survey,t="uuidDisplay";o.Serializer.findProperty("",t)||(o.Serializer.addProperty("",{name:t,type:"string",isSerializable:!1,editor:{render:function(o,t){t.innerHTML="";var r=document.createElement("div");r.style.margin="4px 0",r.style.display="flex",r.style.alignItems="center";var i=document.createElement("div");i.style.display="inline-block",i.style.width="calc(100% - 40px)",i.style.marginRight="8px",i.style.padding="0.25rem 0.5rem",i.style.border="1px solid #dddbda",i.style.borderRadius="4px",i.style.fontSize="0.875rem",i.style.color="#706e6b",i.style.overflow="hidden",i.style.textOverflow="ellipsis",i.style.whiteSpace="nowrap",i.innerText=o.koValue()||"";var l=document.createElement("button");return l.style.minWidth="32px",l.style.height="32px",l.style.border="none",l.style.borderRadius="4px",l.style.backgroundColor="transparent",l.style.cursor="pointer",l.style.padding="4px",l.style.display="flex",l.style.alignItems="center",l.style.justifyContent="center",l.title="Copy to clipboard",l.innerHTML='<svg viewBox="0 0 24 24" width="20" height="20" style="fill: currentColor;">\n                <path d="M16 1H4C3 1 2 2 2 3v14h2V3h12V1zm3 4H8C7 5 6 6 6 7v14c0 1 1 2 2 2h11c1 0 2-1 2-2V7c0-1-1-2-2-2zm0 16H8V7h11v14z"></path>\n            </svg>',l.addEventListener("click",(function(t){t.preventDefault(),t.stopPropagation();var r=o.koValue();r&&e.copyToClipboard(r,this)})),r.appendChild(i),r.appendChild(l),t.appendChild(r),console.log("UUID Display editor rendered",o.koValue()),t}}}),console.log("Registered custom property editor: ".concat(t))),this.registerUuidProperty(o)}else console.error("Survey global object not found")},e.registerUuidProperty=function(e){e.Serializer.addProperty("question",{name:"uuid",displayName:"Question Id",category:"general",visibleIndex:0,type:"uuidDisplay",readOnly:!0}),console.log("Added uuid property to questions using uuidDisplay editor type")},e.copyToClipboard=function(e,o){var t=this;try{navigator.clipboard.writeText(e).then((function(){o.style.color="#2e844a",setTimeout((function(){o.style.color=""}),1e3),console.log("UUID copied to clipboard:",e)})).catch((function(r){console.error("Clipboard copy failed:",r),t.fallbackCopyToClipboard(e,o)}))}catch(t){console.error("Clipboard API error:",t),this.fallbackCopyToClipboard(e,o)}},e.fallbackCopyToClipboard=function(e,o){try{var t=document.createElement("textarea");t.value=e,t.style.position="fixed",document.body.appendChild(t),t.focus(),t.select();var r=document.execCommand("copy");document.body.removeChild(t),r?(console.log("UUID copied via fallback method"),o.style.color="#2e844a",setTimeout((function(){o.style.color=""}),1e3)):console.error("Fallback clipboard copy failed")}catch(e){console.error("Fallback clipboard copy also failed:",e)}},e}();return window.UuidPropertyEditor=o,e=e.default})()));
