--- surveyjs-extensions/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "surveyjs-extensions",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "webpack --mode production",
-    "build:watch": "webpack --mode development --watch",
-    "copy-extensions": "node ./scripts/copy-to-static-resources.js",
-    "build:sf": "npm run build && npm run copy-extensions"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "survey-core": "^2.0.9",
-    "survey-creator-core": "^2.0.9",
-    "ts-loader": "^9.5.2",
-    "typescript": "^5.8.3",
-    "webpack": "^5.99.8",
-    "webpack-cli": "^6.0.1",
-    "terser-webpack-plugin": "^5.3.10"
-  }
-}
